requirements.txt
certifi==2020.12.5
click==7.1.2
Flask==2.0.3
Flask-Session==0.4.0
gunicorn==20.1.0
itsdangerous==2.1.1
MarkupSafe==2.1.0
Werkzeug==2.0.3
mysql-connector-python==8.0.28